pyproject.toml
[project]
name = "acm-2025"
version = "0.1.0"
description = "Instagram scraping and RAG system for ACM hackathon"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Code formatting and quality
    "black>=25.1.0",
    "isort>=6.0.1",
    # LangChain ecosystem
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.22",
    "langchain-ollama>=0.3.6",
    "langchain-openai>=0.3.30",
    # Data processing
    "lxml>=6.0.2",
    "pydantic>=2.12.0",
    "python-dotenv>=1.1.1",
    # Instagram API and web scraping
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    # AI and ML for RAG
    "anthropic>=0.25.0",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "chromadb>=0.4.0",
    # Data analysis
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "pinecone>=7.3.0",
    "streamlit>=1.50.0",
    "google-api-python-client>=2.184.0",
    "google-auth>=2.41.1",
    "google-auth-oauthlib>=1.2.2",
    "google-auth-httplib2>=0.2.0",
    "fastapi>=0.119.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
# RAG System variants
simple-rag = [
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.0",
]
chromadb-rag = [
    "chromadb>=0.4.0",
    "langchain-chroma>=0.1.0",
]
pinecone-rag = [
    "pinecone>=5.0.0",
    "langchain-pinecone>=0.1.0",
]
# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]
# All RAG features
all-rag = [
    "acm-2025[simple-rag,chromadb-rag,pinecone-rag]",
]
